TwitchPlaysPokemon/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
apparmor==3.1.7
attrs==23.2.0
autocommand==2.2.2
Brotli==1.1.0
btrfsutil==6.7.1
cchardet==2.1.7
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
chardet==4.0.0
charset-normalizer==3.3.2
ciso8601==2.3.1
cryptography==42.0.5
distlib==0.3.8
docutils==0.20.1
fastjsonschema==2.19.1
filelock==3.13.1
frozenlist==1.4.1
identify==2.5.35
idna==2.10
inflect==7.0.0
iso8601==2.1.0
jaraco.context==4.3.0
jaraco.functools==4.0.0
jaraco.text==3.12.0
Jinja2==3.0.1
LibAppArmor==3.1.7
MarkupSafe==2.1.5
more-itertools==10.2.0
multidict==6.0.5
mutagen==1.47.0
nodeenv==1.8.0
numpy==1.26.4
opencv-python==4.9.0.80
ordered-set==4.1.0
packaging==23.2
Pillow==8.4.0
platformdirs==4.2.0
ply==3.11
pre_commit==3.6.2
psutil==5.9.8
PyAudio==0.2.11
pycairo==1.26.0
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pygdbmi==0.11.0.0
PyGObject==3.46.0
python-dotenv==1.0.1
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
Reflector==2023.6.28.0.36.1
requests==2.31.0
sentry-sdk==1.42.0
smbus==1.1
toml==0.10.2
tomli==2.0.1
trove-classifiers==2024.3.6
twitchio==2.9.1
typing_extensions==4.10.0
ujson==5.9.0
urllib3==1.26.18
validate-pyproject==0.13.post1.dev0+gb752273.d20230520
virtualenv==20.25.0
websockets==12.0
wget==3.2
yarl==1.9.4
youtube-dl==2021.12.17
yt-dlp==2024.4.9
